Move _mesh_surface from mesher.py to geometry.py
