+++ voltrace/mesher.py
@@ -17,8 +17,6 @@ from . import backend as B
 from .typing import *
 
 __pdoc__ = {}
-__pdoc__['Points3DWithQuads'] = False
-__pdoc__['PointStack'] = False
 __pdoc__['Mesh.__add__'] = True
 
 
@@ -853,282 +851,4 @@ def _ensure_line_orientation(lines: Lines, points: Points3D, should_be_outwards:
             else:
                 p0, p1 = line
                 lines[i] = [p1, p0]
-
-
-class Points3DWithQuads:
-    def __init__(self, indices: ArrayLikeInt1D, quads: QuadsLike) -> None:
-        indices = np.array(indices, dtype=np.int64)
-        quads = np.array(quads, dtype=np.int64)
-        N = len(indices)
-        assert indices.shape == (N, N)
-        assert np.all(quads[:, 1] < N)
-        assert quads.shape == (len(quads), 5)
-        assert np.all(quads[:, 0] == quads[0, 0])
-        
-        self.indices = indices
-        self.quads = quads
-        self.depth = quads[0, 0]
-        
-        self.shape = indices.shape
-    
-    def to_triangles(self) -> Triangles:
-        triangles = []
-
-        def add_triangle(p0, p1, p2):
-            triangles.append([self.indices[p0[0], p0[1]], self.indices[p1[0], p1[1]], self.indices[p2[0], p2[1]]])
-         
-        for quad in self.quads:
-            depth, i0, i1, j0, j1 = quad 
-            assert depth == self.depth
-
-            p0 = (i0, j0)
-            p1 = (i0, j1)
-            p2 = (i1, j1)
-            p3 = (i1, j0)
-
-            split_edge = False
-            
-            # Check if there is a point on the edge 
-            for edge in range(4):
-                # Is there a point on the first edge?
-                point_on_edge = (p0[0]+p1[0])//2, (p0[1]+p1[1])//2
-                
-                if (abs(p0[0] - p1[0]) > 1 or abs(p0[1] - p1[1]) > 1) and \
-                        self.indices[point_on_edge[0], point_on_edge[1]] != -1:
-                    # Yes there is a point.. we have to split the
-                    # quad into three triangles
-                    add_triangle(p0, point_on_edge, p3)
-                    add_triangle(point_on_edge, p2, p3)
-                    add_triangle(point_on_edge, p1, p2)
-                    split_edge = True
-                    break
-                
-                # Rotate the points so we check the next edge
-                p0, p1, p2, p3 = p1, p2, p3, p0
-             
-            if not split_edge: 
-                add_triangle(p0, p1, p2)
-                add_triangle(p0, p2, p3)
-         
-        assert not (-1 in np.array(triangles))
-        return np.array(triangles, dtype=np.uint64)
-    
-    def __getitem__(self, *args: Any, **kwargs: Any) -> ArrayInt1D:
-        return self.indices.__getitem__(*args, **kwargs)
-    
-    def __setitem__(self, *args: Any, **kwargs: Any) -> None:
-        self.indices.__setitem__(*args, **kwargs)
-
-
-class PointStack:
-    def __init__(self, surface: Surface, points: list[Point3D] | None = None) -> None:
-
-        self.points = points if points is not None else [] # here we need points as list to append in-place
-            
-        self.path_length1 = surface.path_length1
-        self.path_length2 = surface.path_length2
-        
-        self.surf = surface
-         
-        self.indices: list[ArrayInt2D] = []
-    
-    def index_to_u(self, depth: int, i: int) -> float:
-        return self.path_length1/(self.get_number_of_indices(depth) - 1) * i
-     
-    def index_to_v(self, depth: int, j: int) -> float:
-        return self.path_length2/(self.get_number_of_indices(depth) - 1) * j
-    
-    def index_to_point(self, depth: int, i: int, j: int) -> Point3D:
-        u = self.index_to_u(depth, i)
-        v = self.index_to_v(depth, j)
-        return self.surf(u, v)
-    
-    def get_number_of_indices(self, depth: int) -> int:
-        return 2**depth + 1
-    
-    def depth(self) -> int:
-        return len(self.indices) - 1
-    
-    def add_level(self) -> None:
-        new_depth = len(self.indices)
-        Nu = Nv = self.get_number_of_indices(new_depth)
-        
-        index_map = np.full((Nu, Nv), -1, dtype=np.int64)
-        
-        if new_depth != 0.:
-            index_map[::2, ::2] = self.indices[-1]
-        
-        self.indices.append(index_map)
-     
-    def to_point_index(self, depth: int, i: int, j) -> int:
-        assert 0 <= i <= self.get_number_of_indices(depth)
-        assert 0 <= j <= self.get_number_of_indices(depth)
-        
-        while depth >= len(self.indices):
-            self.add_level()
-        
-        map_ = self.indices[depth]
-        
-        if map_[i, j] == -1:
-            self.points.append(self.index_to_point(depth, i, j))
-            map_[i, j] = len(self.points) - 1
-
-        return map_[i, j]
-     
-    def __getitem__(self, args: Sequence[int]) -> Points3D:
-        depth, i, j = args
-        return np.array(self.points[self.to_point_index(depth, i, j)])
-     
-    def normalize_to_depth(self, depth: int, quads: QuadsLike, start_depth: int) -> Points3DWithQuads:
-        N = self.get_number_of_indices(depth)
-        
-        while self.depth() < depth:
-            self.add_level()
-        
-        assert self.depth() == depth
-        assert self.indices[-1].shape == (N, N)
-
-        for d in range(start_depth, len(self.indices)-1):
-            previous = self.indices[d]
-            x, y = np.where(previous != -1)
-            self.indices[d+1][2*x, 2*y] = previous[x, y]
-
-        quads = np.array(quads)
-        assert quads.shape == (len(quads), 5)
-        
-        for i in range(len(quads)):
-            quad_depth, i0, i1, j0, j1 = quads[i]
-             
-            while quad_depth < depth:
-                i0 *= 2
-                i1 *= 2
-                j0 *= 2
-                j1 *= 2
-                quad_depth += 1
-              
-            quads[i] = (quad_depth, i0, i1, j0, j1)
-         
-        return Points3DWithQuads(self.indices[-1], quads)
-
-
-def _subdivide_quads(pstack: PointStack, 
-                     mesh_size: float, 
-                     to_subdivide: list[QuadLike] | None = None, 
-                     quads: list[QuadLike] | None = None) -> None: 
-    assert isinstance(pstack, PointStack)
-    
-    to_subdivide = to_subdivide if to_subdivide is not None else []
-    quads = quads if quads is not None else [] # need quads as shared list in order to append in place
-    if not callable(mesh_size):
-        mesh_size_fun = lambda x, y, z: mesh_size
-    else:
-        mesh_size_fun = mesh_size
-
-    while len(to_subdivide) > 0:
-        depth, i0, i1, j0, j1 = to_subdivide.pop()
-        
-        # Determine whether should split horizontally/vertically
-        p1x, p1y, p1z = pstack[depth, i0, j0]
-        p2x, p2y, p2z = pstack[depth, i0, j1]
-        p3x, p3y, p3z = pstack[depth, i1, j0]
-        p4x, p4y, p4z = pstack[depth, i1, j1]
-            
-        horizontal = max(sqrt((p1x-p2x)**2 + (p1y-p2y)**2 + (p1z-p2z)**2), sqrt((p3x-p4x)**2 + (p3y-p4y)**2 + (p3z-p4z)**2))
-        vertical = max(sqrt((p1x-p3x)**2 + (p1y-p3y)**2 + (p1z-p3z)**2) , sqrt((p2x-p4x)**2 + (p2y-p4y)**2 + (p2z-p4z)**2))
-    
-        ms: float = mesh_size_fun((p1x+p2x+p3x+p4x)/4, (p1y+p2y+p3y+p4y)/4, (p1z+p2z+p3z+p4z)/4) # type: ignore
-            
-        h = horizontal > ms or (horizontal > 2.5*vertical and horizontal > 1/8*ms)
-        v = vertical > ms or (vertical > 2.5*horizontal and vertical > 1/8*ms)
-            
-        if h and v: # Split both horizontally and vertically
-            to_subdivide.append((depth+1, 2*i0, 2*i0+1, 2*j0, 2*j0+1))
-            to_subdivide.append((depth+1, 2*i0, 2*i0+1, 2*j0+1, 2*j0+2))
-            to_subdivide.append((depth+1, 2*i0+1, 2*i0+2, 2*j0, 2*j0+1))
-            to_subdivide.append((depth+1, 2*i0+1, 2*i0+2, 2*j0+1, 2*j0+2))
-        elif h and not v: # Split only horizontally
-            to_subdivide.append((depth+1, 2*i0, 2*i1, 2*j0, 2*j0+1))
-            to_subdivide.append((depth+1, 2*i0, 2*i1, 2*j0+1, 2*j0+2)) 
-        elif v and not h: # Split only vertically
-            to_subdivide.append((depth+1, 2*i0, 2*i0+1, 2*j0, 2*j1))
-            to_subdivide.append((depth+1, 2*i0+1, 2*i0+2, 2*j0, 2*j1))
-        else: # We are done, both sides are within mesh size limits
-            quads.append((depth, i0, i1, j0, j1))
-
-def _mesh_subsections_to_quads(surface: Surface, mesh_size: float, start_depth: int) -> tuple[list[Point3D], list[PointStack], list[QuadsLike]]:
-    all_pstacks = []
-    all_quads: list[QuadsLike] = []
-    points: list[Point3D] = []
-    
-    for s in surface._sections():
-        quads: List[QuadLike] = []
-        pstack = PointStack(s, points=points)
-        
-        for i in range(pstack.get_number_of_indices(start_depth) - 1):
-            for j in range(pstack.get_number_of_indices(start_depth) - 1):
-                _subdivide_quads(pstack, mesh_size, to_subdivide=[(start_depth, i, i+1, j, j+1)], quads=quads)
-        
-        all_pstacks.append(pstack)
-        all_quads.append(quads)
-        points = pstack.points
-
-    return points, all_pstacks, all_quads
-    
-def _copy_over_edge(e1: ArrayInt1D, e2: ArrayInt1D) -> None:
-    assert e1.shape == e2.shape
-    mask = e2 != -1
-    e1[mask] = e2[mask]
-    
-    mask = e1 != -1
-    e2[mask] = e1[mask]
-
-def _mesh_surface(surface: Surface, 
-          mesh_size: float, 
-          start_depth: int = 2, 
-          name: str | None = None, 
-          ensure_outward_normals: bool = True) -> Mesh:
-    # Create a point stack for each subsection
-    points_list, point_stacks, quads = _mesh_subsections_to_quads(surface, mesh_size, start_depth)
-     
-    max_depth = max([p.depth() for p in point_stacks])
-     
-    # Normalize all the point stacks to the max depth of all sections 
-    point_with_quads = [p.normalize_to_depth(max_depth, q, start_depth) for p, q in zip(point_stacks, quads)]
-    
-    # Copy over the edges
-    Nx, Ny = len(surface.breakpoints1)+1, len(surface.breakpoints2)+1
-    assert len(point_with_quads) == Nx*Ny
-
-    for i in range(Nx-1):
-        for j in range(Ny): # Horizontal copying
-            _copy_over_edge(point_with_quads[j*Nx + i][-1, :], point_with_quads[j*Nx + i + 1][0, :])
-     
-    for i in range(Nx):
-        for j in range(Ny-1): # Vertical copying
-            _copy_over_edge(point_with_quads[j*Nx + i][:, -1], point_with_quads[(j+1)*Nx + i][:, 0])
-     
-    points = np.array(points_list)
-    triangles = np.concatenate([pq.to_triangles() for pq in point_with_quads], axis=0)
-    
-    assert points.shape == (len(points), 3)
-    assert triangles.shape == (len(triangles), 3)
-    assert np.all( (0 <= triangles) & (triangles < len(points)) )
-     
-    if name is not None:
-        physical_to_triangles = {name:np.arange(len(triangles))}
-    else:
-        physical_to_triangles = {}
-    
-    return Mesh(points=points, triangles=triangles, physical_to_triangles=physical_to_triangles, ensure_outward_normals=ensure_outward_normals)
-
-
-
-
-
-
-
-
-
-
-
